Refactor get account info.

+++ core/communication/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/communication/MetisService.kt
@@ -82,7 +82,7 @@ interface MetisService {
      */
     suspend fun deleteReaction(
         context: MetisContext,
-        reactionId: Int,
+        reactionId: Long,
         serverUrl: String,
         authToken: String
     ): NetworkResponse<Unit>
