+++ core/datastore/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/datastore/MetisStorageService.kt
@@ -45,7 +45,7 @@ interface MetisStorageService {
 
     fun getStoredPosts(
         serverId: String,
-        clientId: Int,
+        clientId: Long,
         filter: List<MetisFilter>,
         sortingStrategy: MetisSortingStrategy,
         query: String?,
