+++ core/datastore/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/datastore/dao/MetisDao.kt
@@ -167,7 +167,7 @@ interface MetisDao {
         courseId: Long,
         exerciseId: Long,
         lectureId: Long,
-        clientId: Int,
+        clientId: Long,
         metisFilter: List<MetisFilter>,
         metisSortingStrategy: MetisSortingStrategy,
         query: String?
