+++ core/datastore/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/datastore/impl/MetisStorageServiceImpl.kt
@@ -382,7 +382,7 @@ class MetisStorageServiceImpl(
 
     override fun getStoredPosts(
         serverId: String,
-        clientId: Int,
+        clientId: Long,
         filter: List<MetisFilter>,
         sortingStrategy: MetisSortingStrategy,
         query: String?,
