+++ core/datastore/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/datastore/room/model/metis/PostReactionEntity.kt
@@ -29,7 +29,7 @@ data class PostReactionEntity(
     @ColumnInfo(name = "emoji")
     val emojiId: String,
     @ColumnInfo(name = "author_id")
-    val authorId: Int,
+    val authorId: Long,
     @ColumnInfo(name = "id")
     val id: Int
 )
